extraction_load/requirements.txt
psycopg2-binary
pandas
python-dotenv
requests
